emptying __init__, placing cleaning scripts in separate file
